+++ depeche_db/tools/pydantic_message_serializer.py
@@ -2,13 +2,38 @@ import datetime as _dt
 import decimal as _decimal
 import enum as _enum
 import uuid as _uuid
-from typing import Any, Optional, Type, TypeVar, no_type_check
+from typing import Any, Callable, Optional, Type, TypeVar, no_type_check
+
+import pydantic as _pydantic
 
 from depeche_db._compat import get_union_members, issubclass_with_union
 
 from .._interfaces import MessageSerializer
 
-T = TypeVar("T")
+pydantic_v2 = _pydantic.VERSION.split(".")[0] == "2"
+T = TypeVar("T", bound=_pydantic.BaseModel)
+
+
+def _get_de_serializer(
+    type_: Type[T],
+) -> tuple[Callable[[dict], T], Callable[[T], dict]]:
+    if pydantic_v2:
+
+        def deserializer_v2(message: dict) -> T:
+            return type_.model_validate(message)
+
+        def serializer_v2(message: T) -> dict:
+            return message.model_dump(mode="json")
+
+        return deserializer_v2, serializer_v2
+
+    def deserializer(message: dict) -> T:
+        return type_.parse_obj(message)
+
+    def serializer(message: T) -> dict:
+        return make_jsonable(message.dict())  # type: ignore[no-any-return]
+
+    return deserializer, serializer
 
 
 class PydanticMessageSerializer(MessageSerializer[T]):
@@ -17,19 +42,20 @@ class PydanticMessageSerializer(MessageSerializer[T]):
     def __init__(
         self, message_type: Any, aliases: Optional[dict[str, Type]] = None
     ) -> None:
-        import pydantic
-
-        version = pydantic.VERSION.split(".")[0]
-        self.pydantic_v2 = version == "2"
-
         self.message_type: type[T] = message_type
-        self.message_types: dict[str, Type] = {}
+        self.message_types: set[str] = set()
+        self.serializers: dict[str, Callable[[Any], dict]] = {}
+        self.deserializers: dict[str, Callable[[dict], Any]] = {}
         for member in get_union_members(message_type):
             if member.__name__ in self.message_types:
                 raise ValueError(
                     f"Duplicate class name {member.__name__} in union {message_type}"
                 )
-            self.message_types[member.__name__] = member
+            self.message_types.add(member.__name__)
+            (
+                self.deserializers[member.__name__],
+                self.serializers[member.__name__],
+            ) = _get_de_serializer(member)
 
         if aliases is not None:
             for alias, member in aliases.items():
@@ -41,7 +67,10 @@ class PydanticMessageSerializer(MessageSerializer[T]):
                     raise ValueError(
                         f"Type {member} is not a subclass of {self.message_type}"
                     )
-                self.message_types[alias] = member
+                self.message_types.add(alias)
+                self.deserializers[alias], self.serializers[alias] = _get_de_serializer(
+                    member
+                )
 
     def serialize(self, message: T) -> dict:
         typename = message.__class__.__name__
@@ -49,30 +78,18 @@ class PydanticMessageSerializer(MessageSerializer[T]):
             raise TypeError(
                 f"Unknown message type {typename} (expected one of {self.message_types})"
             )
-        result: dict = self._serialize(message)
+        result: dict = self.serializers[typename](message)
         result["__typename__"] = typename
         return result
 
     @no_type_check
-    def _serialize(self, message: T) -> dict:
-        if self.pydantic_v2:
-            return message.model_dump(mode="json")
-        return make_jsonable(message.dict())
-
-    @no_type_check
     def deserialize(self, message: dict) -> T:
-        import pydantic as _pydantic
-
         typename = message.pop("__typename__")
         if typename not in self.message_types:
             raise TypeError(
                 f"Unknown message type {typename} (expected one of {self.message_types})"
             )
-        message_type = self.message_types[typename]
-
-        if self.pydantic_v2:
-            return _pydantic.TypeAdapter(message_type).validate_python(message)
-        return _pydantic.parse_obj_as(message_type, message)
+        return self.deserializers[typename](message)
 
 
 def make_jsonable(obj):