+++ docs/CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.10.0
+
+* Add batched acknowledgement strategy
+* Performance improvements
+  * Use less connections on load in subscription
+  * Removed overly defensive check in acknowledgement operation
+  * Pre-calculate hot SqlA queries
+  * Improved Performance in pydantic (de)serialization
+
 # 0.9.1
 
 * Fix passing of batch size in stream/subscription factory